fix py2 long/float/int foo

+++ sourmash_lib/test_sourmash.py
@@ -249,7 +249,7 @@ def test_do_sourmash_compute_with_scaled():
 
         max_hashes = [ x.estimator.max_hash for x in siglist ]
         assert len(max_hashes) == 2
-        assert set(max_hashes) == set([ 2**64/100 ])
+        assert set(max_hashes) == set([ int(2**64 /100.) ])
 
 
 def test_do_sourmash_compute_with_seed():
